requirements.txt
flask
gunicorn
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp